requirements.txt
duckdb
pandas
pyarrow
kagglehub
python-dotenv
dbt-duckdb
